Add files via upload
this is the django project that I have created. It has the apps for the input page(which renders something at the url), the login page, the division view, and the calendar view. *andrew (aka paul) you'll be using models i believe to get the databases set up, and there is a section in the scheduler/scheduler/settings.py file that focuses on the type/platform of databases that you should look at. *elan, you'll be doing things in the scheduler/[app]/views.py files and the static directory and the templates directory; if you want to work together the next session to integrate your html files and the django things, I think we could have a couple different urls supported, at least by the developer server, by the end of the week, which means we would only need to hookup the database and the algorithm to have a mostly functioning, but maybe not pretty website!